backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.12.5
python-multipart==0.0.20
httpx==0.28.1
python-dotenv==1.0.1
google-generativeai==0.8.5
